requirements.txt
mysql-connector-python
devtools